#87 - feat : 마이페이지 개발_북마크 게시글 조회

+++ src/main/java/com/smile/petpat/user/dto/ProfileDto.java
@@ -39,7 +39,7 @@ public class ProfileDto {
         private Long QnaId;
         private String title;
         private LocalDateTime createdAt;
-        private int viewCnt;
+        private Long viewCnt;
         private int commentCnt;
     }
